+++ Wabbajack.App.Wpf/Views/BrowserView.xaml.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Windows.Controls;
+using Microsoft.Web.WebView2.WinForms;
 using ReactiveUI;
 
 namespace Wabbajack.Views;