+++ Wabbajack.DTOs/Game/Game.cs
@@ -52,4 +52,7 @@ public enum Game
     [Description("Mount & Blade II: Bannerlord")] MountAndBlade2Bannerlord,
     [Description("Valheim")]Valheim,
     [Description("Modding Tools")] ModdingTools,
+
+    [Description("Final Fantasy VII Remake")] FinalFantasy7Remake,
+    [Description("Baldur's Gate 3")] BadlursGate3
 }