+++ Wabbajack.Launcher/Views/MainWindow.axaml.cs
@@ -4,7 +4,7 @@ using Avalonia.Markup.Xaml;
 
 namespace Wabbajack.Launcher.Views;
 
-public class MainWindow : Window
+public partial class MainWindow : Window
 {
     public MainWindow()
     {