+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ### Changelog
 
-#### Version - 3.3.1.0 - TBA
+#### Version - 3.4.1.0 - 12/21/2023
+* Added Support for Final Fantasy 7: Remake Intergrade
+* Update CLI to .NET 8.0 (was missed in the last update)
+* Added Support for Baldur's Gate 3 
+  * Very Work in Progress
+  * **NOT** Plug and Play for compiling and installing!
+
+#### Version - 3.4.0.0 - 11/19/2023
 * Fixed `--outputPath` not being used for the CLI `compile` (thanks to @majcosta for fixing that)
 * Improved Log message for cases where low storage on the drive Wabbajack is installed on causes compiles to fail
   * **To list authors still compiling on Wabbajack 3.0.5.0:**  
@@ -11,6 +18,8 @@
 * Updated Dependencies
   * LZ4 to version 1.3.7-beta
   * SharpZipLib to version 1.4.2
+* Fully upgraded to .NET 8.0
+* Fixed the WebView2 based browser window to no longer crash the application when closed
 
 #### Version - 3.3.0.1 - (Was only a Pre-Release)
 * Fixed Manual Downloading on NexusMods being blocked by a hidden cookie consent banner